% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rWeight.R
\name{add_rWeight}
\alias{add_rWeight}
\title{Add reciprocated weights as attribute}
\usage{
add_rWeight(g)
}
\arguments{
\item{g}{a directed weighted graph (object of class igraph)}
}
\value{
a directed weighted graph (object of class igraph) with attribute r_weight
}
\description{
Computes and adds the reciprocated weights as edge attribute to a graph.
}
\examples{
# create a directed weighted graph
library(igraph)
g1 <- graph(c(1,2,2,1, 1,3,3,2), directed=TRUE)
E(g1)$weight <- c(2, 7, 3, 4)
V(g1)$name <- c("a", "b", "c")

# compute/add reciprocated weights
g1 <- add_rWeight(g1)

}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}
\references{
Tiziano Squartini, Francesco Picciolo, Franco Ruzzenenti, Diego Garlaschelli, 
  "Reciprocity of weighted networks", Scientific Reports 3: 2729 (2013).
}

