% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartNetwork.R
\name{bipartNetwork}
\alias{bipartNetwork}
\title{Build bipartite network}
\usage{
bipartNetwork(m)
}
\arguments{
\item{m}{a matrix with one type of observation (e.g., legislators) in rows and the other type of observation (e.g., bills) in columns (see Details)}
}
\value{
a bipartite network (object of class igraph)
}
\description{
Returns a bipartite network (e.g, of bill sponsorship)
}
\details{
Example of bill (co-)sponsorthip: m a matrix with legislators in rows and bills in columns. 
Cell values in m are equal to 1 if legislator i sponsored bill j.
}
\examples{

# graph example
library(igraph)
m <- random_Matrix("binary", 10) 
g <- bipartNetwork(m)
plot(g)

}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}

