% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rwEdges.R
\name{get_rwEdges}
\alias{get_rwEdges}
\title{Get ids of edges that have a reciprocated component}
\usage{
get_rwEdges(g, recipro=TRUE)
}
\arguments{
\item{g}{a directed weighted graph (object of class igraph)}

\item{recipro}{logical, if TRUE (default) only reciprocated compontent is returned, if FALSE only non-reciprocated component is returned}
}
\value{
a numeric vector conteining the respective edge ids
}
\description{
Returns the ids of edges that have a (non-)reciprocated component
}
\examples{

# graph example
library(igraph)
g1 <- graph(c(1,2,2,1, 1,3,3,2), directed=TRUE)
E(g1)$weight <- c(2, 7, 3, 4)
V(g1)$name <- c("a", "b", "c")

# compute/add reciprocated weights
ids <- get_rwEdges(g1)

}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}
\references{
Tiziano Squartini, Francesco Picciolo, Franco Ruzzenenti, Diego Garlaschelli, 
"Reciprocity of weighted networks", Scientific Reports 3: 2729 (2013).
}

