% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weight.R
\name{get_weight}
\alias{get_weight}
\title{Get weight of a given edge}
\usage{
get_weight(g, i)
}
\arguments{
\item{g}{a directed weighted graph (object of class igraph)}

\item{i}{a numeric vector with the ids of edges}
}
\value{
a numeric vector conteining the respective edge weights
}
\description{
Returns the 'weight' attribute of given edges in a directed graph
}
\examples{

# graph example
library(igraph)
g1 <- graph(c(1,2,2,1, 1,3,3,2), directed=TRUE)
E(g1)$weight <- c(2, 7, 3, 4)
V(g1)$name <- c("a", "b", "c")

get_weight(g1,1:2)

}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}

