% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_vertex.R
\name{has_vertex}
\alias{has_vertex}
\title{Find a vertex in a graph}
\usage{
has_vertex(g, v)
}
\arguments{
\item{g}{a directed weighted graph (object of class igraph)}

\item{v}{character string with the name of the vertex}
}
\value{
logical, TRUE if graph g has a vertex with a specific name, FALSE otherwise
}
\description{
Returns TRUE if graph g has a vertex with a specific name
}
\examples{

# graph example
library(igraph)
g1 <- graph(c(1,2,2,1, 1,3,3,2), directed=TRUE)
E(g1)$weight <- c(2, 7, 3, 4)
V(g1)$name <- c("a", "b", "c")

has_vertex(g1, "a")

}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}

