% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jn_sample.R
\name{jn_sample}
\alias{jn_sample}
\title{Jackknife sampling of a directed weighted graph}
\usage{
jn_sample(g)
}
\arguments{
\item{g}{a directed weighted graph (object of class igraph)}
}
\value{
a list of length L (the number of edges in g)
}
\description{
A function that builds a set of jackknife samples of a graph g (removing one edge-weight at a time).
}
\examples{

# graph example
library(igraph)
g1 <- graph(c(1,2,2,1, 1,3,3,2), directed=TRUE)
E(g1)$weight <- c(2, 7, 3, 4)
V(g1)$name <- c("a", "b", "c")

samples <- jn_sample(g1)
length(samples)
E(g1)
samples[[1]]

}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}
\references{
Tiziano Squartini, Franceshttp://127.0.0.1:11315/rstudio/clear.cache.gifco Picciolo, Franco Ruzzenenti, Diego Garlaschelli, 
"Reciprocity of weighted networks", Scientific Reports 3: 2729 (2013).
}

