% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netReciprocity.R
\name{netReciprocity}
\alias{netReciprocity}
\title{Analysis of reciprocity}
\usage{
netReciprocity(g, nm=c("DWCM"), n_samples, ...)
}
\arguments{
\item{g}{a directed weighted graph (object of class igraph) or the respective adjacency matrix (object of class matrix)}

\item{nm}{character, indicating what null model should be used to compute the expected weighted reciprocity (<r>_NM), currently only "DWCM"  implemented.}

\item{n_samples}{numeric, the number of null graphs to be drawn to compute the standard deviation (default is 1500).}

\item{...}{passed down to MAXandSAM()}
}
\value{
a object of class recipro
}
\description{
A function to analyze several aspects of reciprocity in a directed weighted network (see details).
}
\details{
The function computes inter alia the rho reciprocity quantity and the standard deviation of rho based on a sampled network ensemble.
}
\examples{

# directed weighted configuration model example
# (alternatively with the adjacency matrix as input...)
a1 <- random_Matrix(Method="DWCM", n_vertices=10)

netReciprocity(a1, nm="DWCM", 20)

}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}
\references{
Tiziano Squartini, Francesco Picciolo, Franco Ruzzenenti, Diego Garlaschelli, 
"Reciprocity of weighted networks", Scientific Reports 3: 2729 (2013).
}

