% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rw.R
\name{plot_rw}
\alias{plot_rw}
\title{Plot reciprocated graph component}
\usage{
plot_rw(g, extract=TRUE , rcolor="red", othercolor="grey", ...)
}
\arguments{
\item{g}{a directed weighted graph (object of class igraph)}

\item{extract}{logical if TRUE (default), only the reciprocated connections are plotted, if FALSE the whole graph is plotted with reciprocated connections highlighted}

\item{rcolor}{character string, color argument for edges (only used for case extract==FALSE; defaults to "red")}

\item{othercolor}{character string, color argument for non-reciprocated edges (only used for case extract==FALSE; defaults to "grey")}

\item{...}{passed down to igraph}
}
\description{
Plots either only the reciprocated connections of a graph or highlights them in the graph plot.
}
\examples{

# create a directed weighted graph
library(igraph)
g1 <- graph(c(1,2,2,1, 1,3,3,2), directed=TRUE)
E(g1)$weight <- c(2, 7, 3, 4)
V(g1)$name <- c("a", "b", "c")

plot(g1)
plot_rw(g1)
plot_rw(g1, extract=FALSE, edge.arrow.size=0.1, vertex.size=4,
 vertex.label=NA, vertex.color="lightgrey")

}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}

