% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sr.R
\name{plot_sr}
\alias{plot_sr}
\title{Plot reciprocated strength}
\usage{
plot_sr(g)
}
\arguments{
\item{g}{a directed weighted graph (object of class igraph)}
}
\description{
Plots only the vertices of a given graph that have a reciprocated strength > 0.
}
\examples{

# create a directed weighted graph
library(igraph)
g1 <- graph(c(1,2,2,1, 1,3,3,2), directed=TRUE)
E(g1)$weight <- c(2, 7, 3, 4)
V(g1)$name <- c("a", "b", "c")

plot(g1)
plot_sr(g1)

}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}

