% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_dwcm.R
\name{r_dwcm}
\alias{r_dwcm}
\title{Expected weighted reciprocity under DWCM model}
\usage{
r_dwcm(g, ...)
}
\arguments{
\item{g}{a directed weighted graph (object of class igraph) or an adjacency matrix (object of class matrix)}

\item{...}{passed down to MAXandSAM()}
}
\value{
numeric scalar
}
\description{
Computes the expected weighted reciprocity under the directed weighted configuration model (DWCM), given a specific graph g.
}
\examples{

# graph example
library(igraph)
g1 <- graph(c(1,2,2,1, 1,3,3,2), directed=TRUE)
E(g1)$weight <- c(2, 1, 3, 0)
a1 <- random_Matrix(Method="DWCM", n_vertices=10)

# compute r
r_dwcm(g1)
r_dwcm(a1)

}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}
\references{
[1] Tiziano Squartini, Rossana Mastrandrea, Diego Garlaschelli,
"Unbiased sampling of network ensembles", http://arxiv.org/abs/1406.1197 
(Please replace this preprint version with the final published version, as soon as the latter is available.)
[2] Tiziano Squartini, Francesco Picciolo, Franco Ruzzenenti, Diego Garlaschelli, 
"Reciprocity of weighted networks", Scientific Reports 3: 2729 (2013).
}

