% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_Matrix.R
\name{random_Matrix}
\alias{random_Matrix}
\title{Draw a random adjacency matrix}
\usage{
random_Matrix(Method, n_vertices)
}
\arguments{
\item{Method}{a character string indicating the acronym for a specific network model (see details section for a list of models)}

\item{n_vertices}{the number of vertices in the network (dimension of adjacency matrix).}
}
\value{
a matrix (numeric) with dimensions n_vertices x n_vertices.
}
\description{
The function generates a random adjacency matrix in line with a specific network model.
}
\details{
The argument Method needs to refer to one of the following:
'binary' - Random binary matrix
'UBCM' - Undirected Binary Configuration Model
'UWCM' - Undirected Weighted Configuration Model
'DBCM' - Directed Binary Configuration Model
'DWCM'-  Directed Weighted Configuration Model
'UECM'-  Undirected Enhanced Configuration Model
'RBCM'-  Reciprocal Binary Configuration Model
'RWCM'-  Reciprocal Weighted Configuration Model\cr
}
\examples{
W <- random_Matrix("DWCM", 6)
W
}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}
\references{
For details on the different models see:
 Tiziano Squartini, Rossana Mastrandrea, Diego Garlaschelli,
"Unbiased sampling of network ensembles", http://arxiv.org/abs/1406.1197
}

