% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{rho}
\alias{rho}
\title{Roh reciprocity quantity}
\usage{
rho(g, nm="WRG", ...)
}
\arguments{
\item{g}{a directed weighted graph (object of class igraph) or the respective adjacency matrix (object of class matrix)}

\item{nm}{character, indicating what null model should be used to compute the expected weighted reciprocity (<r>_NM), currently either "WRG" or "DWCM".}

\item{...}{passed down to MAXandSAM()}
}
\value{
numeric scalar (between -1 and +1)
}
\description{
Computes the rho reciprocity quantity (comparison of r with r_NM from a null model).
}
\details{
The function is implemented for the follwing null models:
'WRG' -  Weighted Random Graph Model
'DWCM'-  Directed Weighted Configuration Model
}
\examples{

# weighted random graph example
library(igraph)
g1 <- graph(c(1,2,2,1, 1,3,3,2), directed=TRUE)
E(g1)$weight <- c(2, 7, 3, 4)
V(g1)$name <- c("a", "b", "c")

rho(g1, nm="WRG")

# directed weighted configuration model example
# (alternatively with the adjacency matrix as input...)
a1 <- random_Matrix(Method="DWCM", n_vertices=10)

rho(a1, nm="DWCM")

}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}
\references{
Tiziano Squartini, Francesco Picciolo, Franco Ruzzenenti, Diego Garlaschelli, 
"Reciprocity of weighted networks", Scientific Reports 3: 2729 (2013).
}
\seealso{
\code{\link{MAXandSAM}} which this function calls in case of some null models.
}

