% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_sd.R
\name{rho_sd}
\alias{rho_sd}
\title{Roh reciprocity quantity and SD}
\usage{
rho_sd(g, nm=c("DWCM"), n_samples, ...)
}
\arguments{
\item{g}{a directed weighted graph (object of class igraph) or the respective adjacency matrix (object of class matrix)}

\item{nm}{character, indicating what null model should be used to compute the expected weighted reciprocity (<r>_NM), currently only "DWCM"  implemented.}

\item{n_samples}{numeric, the number of null graphs to be drawn to compute the standard deviation (default is 1500).}

\item{...}{passed down to MAXandSAM()}
}
\value{
numeric vector of lenght 2 with rho in the first element and the sd in the second element.
}
\description{
A function that computes the rho reciprocity quantity and the standard deviation of rho based on a sampled network ensemble.
}
\examples{

# directed weighted configuration model example
# (alternatively with the adjacency matrix as input...)
a1 <- random_Matrix(Method="DWCM", n_vertices=10)

rho_sd(a1, nm="DWCM", 10)

}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}
\references{
Tiziano Squartini, Francesco Picciolo, Franco Ruzzenenti, Diego Garlaschelli, 
"Reciprocity of weighted networks", Scientific Reports 3: 2729 (2013).
}

