% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{rw}
\alias{rw}
\title{Reciprocated weight}
\usage{
rw(g, i, j)
}
\arguments{
\item{g}{a directed weighted graph (object of class igraph)}

\item{i}{the id or symbolic vertex name of vertex i}

\item{j}{the id or symbolic vertex name of vertex j}
}
\value{
numeric scalar
}
\description{
Returns the reciprocated weight of the link between vertices i and j.
}
\examples{

# graph example
library(igraph)
g2 <- graph(c(1,2,2,1), directed=TRUE)
E(g2)$weight <- c(2,7)
rw(g2, 1, 2 ) # must be 2
w(g2, 2, 1) - rw(g2, 1, 2 ) # must be 5 (the non-reciprocated weight from 2 to 1 )

}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}
\references{
Tiziano Squartini, Francesco Picciolo, Franco Ruzzenenti, Diego Garlaschelli, 
"Reciprocity of weighted networks", Scientific Reports 3: 2729 (2013).
}

