% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_all.R
\name{rw_all}
\alias{rw_all}
\title{All reciprocated weights}
\usage{
rw_all(g)
}
\arguments{
\item{g}{a directed weighted graph (object of class igraph)}
}
\value{
a dataframe with the edge.ids and the respective reciprocicated weight of each edge
}
\description{
Returns the reciprocated weights between all verteces in a graph.
}
\examples{

# graph example
library(igraph)
g1 <- graph(c(1,2,2,1, 1,3,3,2), directed=TRUE)
E(g1)$weight <- c(2, 7, 3, 4)
V(g1)$name <- c("a", "b", "c")

# compute all reciprocicated weights 
rw_all(g1) 


}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}
\references{
Tiziano Squartini, Francesco Picciolo, Franco Ruzzenenti, Diego Garlaschelli, 
"Reciprocity of weighted networks", Scientific Reports 3: 2729 (2013).
}

