% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snr_in.R
\name{snr_in}
\alias{snr_in}
\title{Non-reciprocicated in-strenght}
\usage{
snr_in(g, i)
}
\arguments{
\item{g}{a directed weighted graph (object of class igraph)}

\item{i}{the id or symbolic vertex name of the vertex to compute the non-reciprocicated in-strength}
}
\value{
a named numeric vector
}
\description{
Returns non-reciprocicated in-strength of vertex i.
}
\examples{

# graph example
library(igraph)
g1 <- graph(c(1,2,2,1, 1,3,3,2), directed=TRUE)
E(g1)$weight <- c(2, 7, 3, 4)
V(g1)$name <- c("a", "b", "c")

# compute the non-reciprocicated in-strenght of a vertex
snr_in(g1, 1)
snr_in(g1, "a") 
snr_in(g1, "b")


}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}
\references{
Tiziano Squartini, Francesco Picciolo, Franco Ruzzenenti, Diego Garlaschelli, 
"Reciprocity of weighted networks", Scientific Reports 3: 2729 (2013).
}

