% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr.R
\name{sr}
\alias{sr}
\title{Reciprocicated strength}
\usage{
sr(g, i)
}
\arguments{
\item{g}{a directed weighted graph (object of class igraph)}

\item{i}{the id or symbolic vertex name of the vertex to compute the reciprocicated strength}
}
\value{
a named numeric vector
}
\description{
Returns the reciprocicated strength of vertex i.
}
\examples{

# graph example
library(igraph)
g2 <- graph(c(1,2,2,1,2,3,3,2), directed=TRUE)
E(g2)$weight <- c(2,7,3,3)

sr(g2, 2) # must be 5

}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}
\references{
Tiziano Squartini, Francesco Picciolo, Franco Ruzzenenti, Diego Garlaschelli, 
"Reciprocity of weighted networks", Scientific Reports 3: 2729 (2013).
}

