% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w.R
\name{w}
\alias{w}
\title{Weight of directed edge}
\usage{
w(g, i, j, eid=NULL)
}
\arguments{
\item{g}{a directed weighted graph (object of class igraph)}

\item{i}{the id or symbolic vertex name of vertex i}

\item{j}{the id or symbolic vertex name of vertex j}

\item{eid}{the id of the edge (can be provided instead of i or j)}
}
\value{
numeric scalar
}
\description{
Returns the weight w.ij of a directed edge from vertex i to vertex j.
}
\examples{

# graph example
library(igraph)
g2 <- graph(c(1,2,2,1), directed=TRUE)
E(g2)$weight <- c(2,7)
w(g2, 2, 1)
w(g2, NULL, NULL, 2) 

}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}

